update to fix a signed integer error

+++ ngx_http_memcached_hash_module.c
@@ -456,7 +456,7 @@ ngx_http_memcached_hash_filter(void *data, ssize_t bytes)
     u = ctx->request->upstream;
     b = &u->buffer;
 
-    if (u->length == ctx->rest) {
+    if ((size_t)u->length == ctx->rest) {
 
         if (ngx_strncmp(b->last,
                    ngx_http_memcached_hash_end + NGX_HTTP_MEMCACHED_HASH_END - ctx->rest,
